Stop treating *.mas and *.mc as HTML::Mason files
There was a time when I worked with HTML::Mason - and edited
HTML::Mason .mas and .mc files.

I'm pretty sure that if for some reason that happens again - and I
feel the need for any syntax highlightning on them, I will remember to
remove this comments, but... nowadays, chances are I am going to hit
this when editing some file with one of thse extensions that is *not*
an HTML::Mason file... At least that's what triggered this commit.